fixed tiny bug in convert (int -> long long).

+++ arch/x86_64/include/utils/convert.h
@@ -1,7 +1,7 @@
 #ifndef UTILS_CONVERT_H
 #define UTILS_CONVERT_H
 
-char* dec(unsigned x, char *s);
-char* decX(unsigned x, char *s);
+char* dec(unsigned long long x, char *s);
+char* decX(unsigned long long x, char *s);
 
 #endif
